ENH: OutSourcing of charts into qSlicerMultiVolumeExplorerCharts.py - reorganized files for better readability - extracted more methods
